BREAKING CHANGE: introduce a new Character type and optionally use it for "char"
The handling of /char has always been wrong in Typelib. It was
interpreted as a numeric type, always of the signed-ness of the
local architecture. This broke (1) display (as we did not know
whether a numeric is a character or not and (2) multiarch
as the signed-ness changes.

This commit finally does the right thing, and makes /char a
character type of size 1.