fix: create one and only one PluginManager in a Ruby process
Typelib::PluginManager is managed as a phoenix singleton, which means
that it is destroyed whenever it is not in use. Within the Ruby
extension, this means that we create one every time we create a
Registry object - and creating one is expensive, given that it
looks for plugins in a PATH every time.

Moreover, it is not thread safe - two C++ registry objects created
more or less at the same time in two different threads could lead to
crashes.